pyproject.toml
[project]
name = "pyhton-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask==3.1.2",
    "flask-cors==5.0.0",
    "protobuf==5.28.3",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "upstox-python-sdk==2.18.0",
    "websockets==15.0.1",
]
